catch to xunit conversion
Unfortunately, there is an arguable bug in Catch's xunit output
(anything larger than 300,000 bytes or so is clipped, resulting
in invalid XML). The CSMP tests are very chatty, so this happens
basically all the time.

The only solution is to convert the Catch XML format into xunit.
This is lossy, but it's the best we can do for now.