chore: refactored H2O_CO2_NaCl_FlowFunctions because they were not mass-based yet.
NOTE: the gravity term was reformulated and now returs a vectorvariable because, in the case of lower dimensional elements
such as inclide wells the term must act in the direction of the well, else the computations are wrong.
This causes a compilation issue in DES which we need to resolve.